# api/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Revues API
  version: 0.1.0
  description: 'Contrat HTTP JSON de Revues (`/api/v1`). Source de vérité pour le codegen

    serveur (oapi-codegen) et le client TypeScript (WP-005).

    '
servers:
- url: /api/v1
  description: API versionnée
...
paths:
...
  /subjects/{subjectId}/members:
    parameters:
    - $ref: '#/components/parameters/SubjectId'
    get:
      operationId: listSubjectMembers
      tags:
      - subjects
      summary: Membres directs du sujet
      description: Auth + CanViewAccess. Liste `subject_members` seulement.
      responses:
        '200':
          description: Membres directs
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubjectMemberListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      operationId: addSubjectMember
      tags:
      - subjects
      summary: Ajouter un membre direct
      description: 'Auth + CanManageSubjectMembers / CanInviteSubjectMember.

        L''invité doit déjà avoir un compte (email). CSRF requis.

        '
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddSubjectMemberRequest'
      responses:
        '201':
          description: Membre ajouté
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubjectMember'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /runs:
    get:
      operationId: listRuns
      tags:
        - runs
      summary: Liste des revues visibles
      description: Auth + org. Filtre status/q. Visibilite via ResolveSubjectAccess.
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum:
              - draft
              - in_progress
              - done
              - overdue
          description: Filtre statut
        - name: q
          in: query
          required: false
          schema:
            type: string
          description: Recherche texte
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 25
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        '200':
          description: Liste
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
  /subjects/{subjectId}/runs:
    parameters:
      - $ref: '#/components/parameters/SubjectId'
    get:
      operationId: listSubjectRuns
      tags:
        - runs
      summary: Revues d'un sujet
      description: Auth + CanViewAccess sur le sujet.
      responses:
        '200':
          description: Liste
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      operationId: createRun
      tags:
        - runs
      summary: Lancer une revue (snapshot transactionnel)
      description: Auth + CanLaunchAccess. Copie SQL template_items vers run_items en transaction. CSRF requis.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRunRequest'
      responses:
        '201':
          description: Revue creee
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunDetail'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /subjects/{subjectId}/run-templates:
    parameters:
      - $ref: '#/components/parameters/SubjectId'
    get:
      operationId: listSubjectRunTemplates
      tags:
        - runs
      summary: Modeles lanables pour un sujet
      description: Auth + CanViewAccess. Matching domaines sujet / modeles.
      responses:
        '200':
          description: Catalogue filtre
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunTemplateListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /runs/{runId}/complete:
    parameters:
      - $ref: '#/components/parameters/RunId'
    post:
      operationId: completeRun
      tags:
        - runs
      summary: Cloturer une revue
      description: Auth + CanCompleteAccess. Bloque si points obligatoires pending. Emit review.completed. CSRF.
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompleteRunRequest'
      responses:
        '200':
          description: Revue cloturee
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunDetail'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalError'
  /runs/{runId}/items/{itemId}:
    parameters:
      - $ref: '#/components/parameters/RunId'
      - $ref: '#/components/parameters/RunItemId'
    get:
      operationId: getRunItem
      tags:
        - runs
      summary: Detail d'un point + audit
      description: Auth + Visible. Inclut historique run_item_events.
      responses:
        '200':
          description: Point
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunItemDetail'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    patch:
      operationId: updateRunItem
      tags:
        - runs
      summary: Mettre a jour status/comment/assign
      description: Auth + CanUpdateAccess; assign exige CanAssignAccess. Comment obligatoire si nok. Run done -> 409. 
        Emit review.item.nok. CSRF.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateRunItemRequest'
      responses:
        '200':
          description: Point mis a jour
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunItemDetail'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalError'
components:
  responses:
    BadRequest:
      description: Requête invalide
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Non authentifié ou identifiants invalides
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: CSRF manquant/invalide ou refus
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Ressource absente ou non visible (IDOR)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: Conflit métier (ex. version publiée immuable)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalError:
      description: Erreur interne
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    SubjectMember:
      type: object
      required:
      - user_id
      - login
      - email
      - display_name
      - role
      properties:
        user_id:
          type: integer
          format: int64
        login:
          type: string
        email:
          type: string
          format: email
        display_name:
          type: string
        role:
          type: string
          enum:
          - lead
          - contributor
          - viewer
    SubjectMemberListResponse:
      type: object
      required:
      - members
      properties:
        members:
          type: array
          items:
            $ref: '#/components/schemas/SubjectMember'
...
    AddSubjectMemberRequest:
      type: object
      required:
      - email
      properties:
        email:
          type: string
          format: email
        role:
          type: string
          enum:
          - lead
          - contributor
          - viewer
          default: viewer
...
    RunListResponse:
      type: object
      required:
        - runs
        - total
      properties:
        runs:
          type: array
          items:
            $ref: '#/components/schemas/RunSummary'
        total:
          type: integer
...
    RunTemplateListResponse:
      type: object
      required:
        - templates
        - can_launch
      properties:
        templates:
          type: array
          items:
            $ref: '#/components/schemas/RunTemplateSummary'
        can_launch:
          type: boolean
    CreateRunRequest:
      type: object
      required:
        - template_id
      properties:
        template_id:
          type: integer
          format: int64
          minimum: 1
        due_date:
          type: string
          description: Date YYYY-MM-DD optionnelle
          nullable: true
    CompleteRunRequest:
      type: object
      properties:
        closing_note:
          type: string
          description: Note de cloture optionnelle
...
    RunItemDetail:
      type: object
      required:
        - item
        - events
        - capabilities
      properties:
        item:
          $ref: '#/components/schemas/RunItem'
        events:
          type: array
          items:
            $ref: '#/components/schemas/RunItemEvent'
        assignees:
          type: array
          items:
            $ref: '#/components/schemas/RunAssignee'
        capabilities:
          $ref: '#/components/schemas/RunCapabilities'
        run_status:
          type: string
        run_title:
          type: string
        subject_id:
          type: integer
          format: int64
        subject_name:
          type: string
...
    RunDetail:
      type: object
      required:
        - id
        - title
        - subject_id
        - subject_name
        - status
        - template_id
        - template_name
        - template_version
        - created_at
        - progress
        - items
        - capabilities
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        subject_id:
          type: integer
          format: int64
        subject_name:
          type: string
        status:
          type: string
          enum:
            - draft
            - in_progress
            - done
            - archived
        due_date:
          type: string
          nullable: true
        closing_note:
          type: string
        template_id:
          type: integer
          format: int64
        template_name:
          type: string
        template_version:
          type: integer
        created_at:
          type: string
        started_at:
          type: string
          nullable: true
        completed_at:
          type: string
          nullable: true
        progress:
          $ref: '#/components/schemas/RunProgress'
        items:
          type: array
          items:
            $ref: '#/components/schemas/RunItem'
        assignees:
          type: array
          items:
            $ref: '#/components/schemas/RunAssignee'
        capabilities:
          $ref: '#/components/schemas/RunCapabilities'
        pending_required_count:
          type: integer
  parameters:
    SubjectId:
      name: subjectId
      in: path
      required: true
      schema:
        type: integer
        format: int64
...
    RunId:
      name: runId
      in: path
      required: true
      schema:
        type: integer
        format: int64
        minimum: 1
      description: Identifiant de la revue
    RunItemId:
      name: itemId
      in: path
      required: true
      schema:
        type: integer
        format: int64
        minimum: 1
      description: Identifiant du point (run item)